fix image URLs on production
Fix image URLs on production. For this all images were moved into the
public folder, since we cannot easily(if even possible) resolve the
asset url of an image residing in the docs, when its a prop in a vue
component.

We use the correct form to link images in the public folder by
prepending the url with a `/`. You will notice this is not done for
those paths int he vue components. There `./` is used. This will prevent
a conflict with a `withBase` defined in vuepress, but those URLs must be
relative to the docs root. In our case its just the `/index.md` hence
`./folder/path/in/public` is implied.